[Services -> Settings] Add custom value support.
