+++ GyroShell.Library/Services/Environment/IDispatcherService.cs
@@ -2,6 +2,9 @@
 
 namespace GyroShell.Library.Services.Environment
 {
+    /// <summary>
+    /// Provides an abstraction of GyroShell's internal dispatcher.
+    /// </summary>
     public interface IDispatcherService
     {
         DispatcherQueue DispatcherQueue { get; init; }
